refactor: Rewrite the main page into 2 logical parts
