+++ src/components/formSearchInput/formSearchInput.css
@@ -1,0 +1,10 @@
+.formSearchRow {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  gap: 10px;
+}
+
+.error {
+  color: red;
+}
